use common component
