+++ libs/designer/src/components/common/mod.rs
@@ -1,0 +1,3 @@
+mod iframe_portal;
+
+pub use iframe_portal::*;
