+++ libs/designer/src/components/mod.rs
@@ -1,1 +1,2 @@
 pub mod main;
+pub mod common;
